Multi image color mod
Strips away the online server, leaving just a script that color batches of images at once.

Files are named like: 000000000000001_ID.png to allow you to put them into frames. The downside to this convenience is that you won't get a preview or be able to select points.

It takes an input (--input), an output (--output) and an optional color palate to apply to all images (--images). The session is not kept alive after finishing, so the startup cost will always be present.