Added PaintLight and Threads
This mod adds in the original painting light, minus the ancient super resolution technique used. It has some hard coded lighting and angles but you can change them however you like. I am not entirely sure what each of the values were supposed to be, but I reversed it enough to get it added in. Just as a note, this is not the same as the one in V4, it's the old ray tracing paper and will look terrible.

The threading was added just to test, it functions but not great.

The prior mod could generate an image in all styles (had to) within 15 seconds. This release, with all settings set, will take about 10 minutes per image.

Added in grading. Grading allows you to set how many times the colors will be refined, from 0 (careless as the OG creator called it) to infinity (not recommended). This can get rid of some of those banding issues in the colors. Speaking of styles, you can also select what kind of style you want this thing to produce as a final output. If you only want flats produced, use 1 for style, or for blended smooth you can use 4. If you set it to 0 it will generate all styles and light all styles but only the final grade.